python-sdk-resource-creation-samples/create-storage-account/hosted/requirements.txt
flask
azure-mgmt-resource
azure-mgmt-storage
azure-identity
